Parser.java: Change to class-level
Several methods of the parser are changed from instance-level to class-level methods as they are not associated with any objects in any way.
With this change, it will not introduce new source of behavior as it will work the same regardless of whether it is declared as a class or instance-level method.